requirements.txt
Flask==3.0.0
keras>=3.10.0
tensorflow>=2.16.1
Pillow==10.2.0
gunicorn==21.2.0